+++ packages/@expo/cli/src/start/server/DevServerManager.ts
@@ -164,6 +164,10 @@ export class DevServerManager {
     await server.waitForTypeScriptAsync();
   }
 
+  async watchEnvironmentVariables() {
+    await devServers.find((server) => server.name === 'metro')?.watchEnvironmentVariables();
+  }
+
   /** Stop all servers including ADB. */
   async stopAsync(): Promise<void> {
     await Promise.allSettled([