+++ packages/@expo/cli/ts-declarations/metro/index.d.ts
@@ -24,70 +24,6 @@ declare module 'metro/src/HmrServer' {
   module.exports = MetroHmrServer;
 }
 
-declare module 'metro/src/ModuleGraph/worker/collectDependencies' {
-  export type AllowOptionalDependenciesWithOptions = {
-    exclude: string[];
-  };
-
-  export type DynamicRequiresBehavior = 'throwAtRuntime' | 'reject';
-
-  export type AllowOptionalDependencies = boolean | AllowOptionalDependenciesWithOptions;
-}
-
-declare module 'metro/src/DeltaBundler/types.flow' {
-  export type AllowOptionalDependenciesWithOptions = {
-    exclude: string[];
-  };
-
-  export type AllowOptionalDependencies = boolean | AllowOptionalDependenciesWithOptions;
-}
-declare module 'metro/src/DeltaBundler' {
-  export type AsyncDependencyType = 'async' | 'prefetch';
-  export type TransformResultDependency = {
-    /**
-     * The literal name provided to a require or import call. For example 'foo' in
-     * case of `require('foo')`.
-     */
-    name: string;
-
-    /**
-     * Extra data returned by the dependency extractor.
-     */
-    data: {
-      /**
-       * A locally unique key for this dependency within the current module.
-       */
-      key: string;
-      /**
-       * If not null, this dependency is due to a dynamic `import()` or `__prefetchImport()` call.
-       */
-      asyncType: AsyncDependencyType | null;
-      /**
-       * The condition for splitting on this dependency edge.
-       */
-      splitCondition?: {
-        mobileConfigName: string;
-      };
-      /**
-       * The dependency is enclosed in a try/catch block.
-       */
-      isOptional?: boolean;
-
-      locs: $ReadOnlyArray<BabelSourceLocation>;
-
-      /** Context for requiring a collection of modules. */
-      contextParams?: RequireContextParams;
-    };
-  };
-}
-
-declare module 'metro/src/lib/countLines' {
-  const countLines = (string: string) => number;
-
-  module.exports = countLines;
-  export default countLines;
-}
-
 declare module 'metro/src/lib/createWebsocketServer' {
   export function createWebsocketServer<TClient extends object>({
     websocketServer,
@@ -96,90 +32,50 @@ declare module 'metro/src/lib/createWebsocketServer' {
   module.exports = createWebsocketServer;
 }
 
-declare module 'metro/src/lib/splitBundleOptions' {
-  import { ConfigT } from 'metro-config';
+declare module 'metro/src/DeltaBundler/Serializers/baseJSBundle' {
+  import { Module, Graph, SerializerOptions } from 'metro';
 
-  type SplitBundleOptions = {
-    entryFile: string;
-    resolverOptions: unknown;
-    transformOptions: { platform: string };
-    serializerOptions: unknown;
-    graphOptions: unknown;
-    onProgress?: (numProcessed: number, total: number) => any;
+  type ModuleMap = readonly [number, string][];
+
+  type Bundle = {
+    readonly modules: ModuleMap;
+    readonly post: string;
+    readonly pre: string;
   };
 
-  type BundleOptions = any;
+  export default function baseJSBundle(
+    entryPoint: string,
+    preModules: readonly Module[],
+    graph: Graph,
+    options: SerializerOptions
+  ): Bundle;
 
-  function splitBundleOptions(options: BundleOptions): SplitBundleOptions;
-
-  export default splitBundleOptions;
+  module.exports = baseJSBundle;
 }
 
-declare module 'metro/src/DeltaBundler/Serializers/helpers/js' {
-  import type { JsOutput } from 'metro-transform-worker';
-  import type { MixedOutput, Module } from 'metro';
+declare module 'metro/src/lib/bundleToString' {
+  import { Module, ReadOnlyGraph, SerializerOptions } from 'metro';
 
-  export function getJsOutput(
-    module: readonly {
-      output: readonly MixedOutput[];
-      path?: string;
-    }
-  ): JsOutput;
+  type ModuleMap = readonly [number, string][];
 
-  export function isJsModule(module: Module<unknown>): boolean;
-}
-
-declare module 'metro/src/Assets' {
-  export type AssetInfo = {
-    files: string[];
-    hash: string;
-    name: string;
-    scales: number[];
-    type: string;
+  type Bundle = {
+    readonly modules: ModuleMap;
+    readonly post: string;
+    readonly pre: string;
   };
 
-  export type AssetDataWithoutFiles = {
-    __packager_asset: boolean;
-    fileSystemLocation: string;
-    hash: string;
-    height: number | null;
-    httpServerLocation: string;
-    name: string;
-    scales: number[];
-    type: string;
-    width: number | null;
+  type BundleMetadata = {
+    readonly pre: number;
+    readonly post: number;
+    readonly modules: readonly [number, number][];
   };
 
-  export type AssetDataFiltered = {
-    __packager_asset: boolean;
-    hash: string;
-    height: number | null;
-    httpServerLocation: string;
-    name: string;
-    scales: number[];
-    type: string;
-    width: number | null;
+  export default function bundleToString(bundle: Bundle): {
+    readonly code: string;
+    readonly metadata: BundleMetadata;
   };
 
-  export type AssetData = AssetDataWithoutFiles & { files: Array<string> };
-
-  export type AssetDataPlugin = (assetData: AssetData) => AssetData | Promise<AssetData>;
-
-  export async function getAsset(
-    relativePath: string,
-    projectRoot: string,
-    watchFolders: readonly string[],
-    platform: string | null | undefined,
-    assetExts: readonly string[]
-  ): Promise<Buffer>;
-
-  async function getAssetData(
-    assetPath: string,
-    localPath: string,
-    assetDataPlugins: readonly string[],
-    platform: string | null | undefined,
-    publicPath: string
-  ): Promise<AssetData>;
+  module.exports = baseJSBundle;
 }
 
 declare module 'metro' {
@@ -202,8 +98,8 @@ declare module 'metro' {
   //#endregion
   //#region metro/src/DeltaBundler/types.flow.js
 
-  export interface MixedOutput {
-    readonly data: unknown;
+  interface MixedOutput {
+    readonly data: any;
     readonly type: string;
   }
 
@@ -257,6 +153,11 @@ declare module 'metro' {
     readonly data: TransformResultDependency;
   }
 
+  export type MixedOutput = {
+    readonly data: any;
+    type: string;
+  };
+
   export interface Module<T = MixedOutput> {
     readonly dependencies: Map<string, Dependency>;
     readonly inverseDependencies: Set<string>;
@@ -269,6 +170,13 @@ declare module 'metro' {
     dependencies: Map<string, Module<T>>;
     importBundleNames: Set<string>;
     readonly entryPoints: ReadonlyArray<string>;
+  }
+
+  export interface ReadOnlyGraph<T = MixedOutput> {
+    readonly entryPoints: Set<string>;
+    // Unused in core but useful for custom serializers / experimentalSerializerHook
+    readonly transformOptions: unknown;
+    readonly dependencies: unknown;
   }
 
   export type TransformResult<T = MixedOutput> = Readonly<{
@@ -318,7 +226,7 @@ declare module 'metro' {
 
   import { Server as HttpServer } from 'http';
   import { Server as HttpsServer } from 'https';
-  import { loadConfig, ConfigT, InputConfigT, Middleware, ConfigT } from 'metro-config';
+  import { loadConfig, ConfigT, InputConfigT, Middleware } from 'metro-config';
 
   type MetroMiddleWare = {
     attachHmrServer: (httpServer: HttpServer | HttpsServer) => void;